additional styling changes
